requirements.txt
bs4
lxml
requests
streamlit
Pillow

